feat: add create_access_token() in auth_dependency.py and test in auth.py
Implemented the create_access_token() method in auth_dependency.py to generate access tokens. Verified its functionality by conducting tests in auth.py.

- auth_dependency.py: Added create_access_token() method for generating access tokens.
- auth.py: Conducted tests to ensure the proper functioning of create_access_token().
